requirements.txt
anyio==3.6.1
autopep8==1.7.0
click==8.1.3
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
mysqlclient==2.1.1
pycodestyle==2.9.1
pydantic==1.10.1
python-dotenv==0.20.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
